+++ src/components/Alert/Alert.tsx
@@ -3,6 +3,7 @@ import Icon from "components/Icon";
 import View from "components/View";
 import Text from "components/Text";
 import * as T from "./Alert.types";
+import s from "./Alert.module.css";
 
 const Alert = (props: T.Props) => {
 	const {
@@ -83,7 +84,9 @@ const Alert = (props: T.Props) => {
 		>
 			{icon ? (
 				<>
-					<Icon svg={icon} size={5} color={isNeutral ? "primary" : color} />
+					<div className={s.icon}>
+						<Icon svg={icon} size={5} color={isNeutral ? "primary" : color} />
+					</div>
 					<View.Item grow>{applyActions(renderContent())}</View.Item>
 				</>
 			) : (